Fix#12 (#14)
* removido os arquivos de spec e solucionado os erros descritos no #12

* modificado o init pra ficar igual ao model

+++ src/document/document.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateDocumentDto } from './dto/create-document.dto';
 import { UpdateDocumentDto } from './dto/update-document.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
@@ -7,38 +7,62 @@ import { PrismaService } from 'src/prisma/prisma.service';
 export class DocumentService {
   constructor(private readonly prisma: PrismaService) {}
 
-  async create(createDocumentDto: CreateDocumentDto) {
+  async createUser(createDocumentDto: CreateDocumentDto) {
     const data = {
-      ...createDocumentDto
+      ...createDocumentDto,
     };
-    await this.prisma.document.create({ data });
+    const createdDocument = await this.prisma.document.create({ data });
+    const { createdAt, updatedAt, ...document } = createdDocument;
+    return document;
+  }
+
+  async updateDocument(id: number, updateDocumentDto: UpdateDocumentDto) {
+    const document = await this.findById(id);
+    if (!document) {
+      throw new NotFoundException('Document not found');
+    }
+    const updatedDocument = await this.prisma.document.update({
+      where: { id },
+      data: updateDocumentDto,
+    });
+    const { createdAt, updatedAt, ...updatedDocumentData } = updatedDocument;
+    return updatedDocumentData;
+  }
+
+  async findDocumentById(id: number) {
+    const document = await this.findById(id);
+    if (!document) {
+      throw new NotFoundException('Document not found');
+    }
+    return await this.prisma.document.findUnique({
+      where: { id },
+    });
+  }
+
+  async findById(id: number) {
+    return await this.prisma.document.findUnique({
+      where: { id },
+    });
   }
 
   async findAll() {
-    return this.prisma.document.findMany({
+    return await this.prisma.document.findMany({
       select: {
         id: true,
-        creatorId: true,
-        content: true
-      }
+        content: true,
+      },
+      orderBy: {
+        id: 'desc',
+      },
     });
   }
 
-  async findOne(id: number) {
-    return this.prisma.document.findUnique({
-      where: { id },
-    });
-  }
-
-  async update(id: number, newData: UpdateDocumentDto) {
-    return this.prisma.document.update({
-      where: { id },
-      data: newData,
-    });
-  }
-
-  async remove(id: number) {
-    return this.prisma.document.delete({
+  async deleteDocument(id: number) {
+    const document = await this.findById(id);
+    if (!document) {
+      throw new NotFoundException('Document not found');
+    }
+    await this.prisma.document.delete({
       where: { id },
     });
   }
